+++ application/config/routes.php
@@ -55,6 +55,7 @@ $route['stockhistory'] = "Stock";
 $route['player'] = "Player";
 $route['404_override'] = '';
 //$route['translate_uri_dashes'] = FALSE;
+$route["stockhistory/mostRecent"] = "Stock/getMostRecentStock";
 $route["stockhistory/(:any)"] = "Stock/getSelectedStock/$1";