+++ application/views/_template.php
@@ -24,20 +24,20 @@ if (!defined('APPPATH'))
 <!--        <link href="/assets/css/bootstrap.min.css" rel="stylesheet" media="screen"/>-->
 <!--        <link rel="stylesheet" type="text/css" href="/assets/css/style.css"/>-->
         <!-- Bootstrap Core CSS -->
-        <link href="assets/css/bootstrap.min.css" rel="stylesheet">
+        <link href="/assets/css/bootstrap.min.css" rel="stylesheet">
 
         <!-- Custom CSS -->
-        <link href="assets/css/style.css" rel="stylesheet">
+        <link href="/assets/css/style.css" rel="stylesheet">
 
-        <link href="assets/css/modal.css" rel="stylesheet">
+        <link href="/assets/css/modal.css" rel="stylesheet">
 
-        <script src="assets/js/modernizr.custom.js"></script>
+        <script src="/assets/js/modernizr.custom.js"></script>
 
         <!-- Morris Charts CSS -->
-        <link href="assets/css/plugins/morris.css" rel="stylesheet">
+        <link href="/assets/css/plugins/morris.css" rel="stylesheet">
 
         <!-- Custom Fonts -->
-        <link href="assets/font-awesome/css/font-awesome.min.css" rel="stylesheet" type="text/css">
+        <link href="/assets/font-awesome/css/font-awesome.min.css" rel="stylesheet" type="text/css">
 
         <!-- Fonts -- >
 <link href='https://fonts.googleapis.com/css?family=Open+Sans:400,600,700,300,400italic,300italic,600italic,700italic,800italic,800' rel='stylesheet' type='text/css'>
@@ -80,7 +80,7 @@ if (!defined('APPPATH'))
         <div id="sidebar-wrapper">
             <ul class="sidebar-nav">
                 <div class = "profile">
-                    <img src="assets/img/profile.png" class = "img-responsive avatar">
+                    <img src="/assets/img/profile.png" class = "img-responsive avatar">
                     <h1 id = "name-header">Alvin Man</h1>
                 </div>
                 <li>
@@ -99,7 +99,7 @@ if (!defined('APPPATH'))
                     <a href="/">Dashboard</a>
                 </li>
                 <li>
-                    <a href="/stockhistory">Stock History</a>
+                    <a href="/stockhistory/mostRecent">Stock History</a>
                 </li>
                 <li>
                     <a href="/player">Player</a>
@@ -116,18 +116,18 @@ if (!defined('APPPATH'))
         {content}
 
         <!-- jQuery -->
-        <script src="assets/js/jquery-1.11.1.min.js"></script>
+        <script src="/assets/js/jquery-1.11.1.min.js"></script>
 
         <!-- Bootstrap Core JavaScript -->
-        <script src="assets/js/bootstrap.min.js"></script>
+        <script src="/assets/js/bootstrap.min.js"></script>
 
-        <script src="assets/js/classie.js"></script>
-        <script src="assets/js/modalEffects.js"></script>
+        <script src="/assets/js/classie.js"></script>
+        <script src="/assets/js/modalEffects.js"></script>
 
         <!-- Morris Charts JavaScript -->
-        <script src="assets/js/plugins/morris/raphael.min.js"></script>
-        <script src="assets/js/plugins/morris/morris.min.js"></script>
-        <script src="assets/js/plugins/morris/morris-data.js"></script>
+        <script src="/assets/js/plugins/morris/raphael.min.js"></script>
+        <script src="/assets/js/plugins/morris/morris.min.js"></script>
+        <script src="/assets/js/plugins/morris/morris-data.js"></script>
 
         <!-- Menu Toggle Script -->
         <script>